Use separate key id/path for jaas (part2).
