+++ tests/env.example
@@ -24,18 +24,25 @@
 # The tenant used when executing the iframeAPI tests, will override any tenant from BASE_URL if any
 #IFRAME_TENANT=
 
-# Configure the domain and tenant for jaas-specific tests (specs/jaas). Note that some of the iFrame tests can also
-# be used to test JAAS, they are configured separately via IFRAME_TENANT.
+# Configure properties for jaas-specific tests (specs/jaas). Note that some of the iFrame tests can also
+# be used to test JaaS, they are configured separately via IFRAME_TENANT, JWT_KID, JWT_PRIVATE_KEY_PATH.
+# Domain for the JaaS environment, e.g. stage.8x8.vc
 JAAS_DOMAIN=
+# The key ID
+JAAS_KID=
+# The path to the private key used for generating JWT token (.pk) for jaas-specific tests
+JAAS_PRIVATE_KEY_PATH=
+# The JaaS tenant, e.g. vpaas-magic-cookie-abcdabcd1234567890
 JAAS_TENANT=
 
 # An access token to use to create meetings (used for the first participant)
 #JWT_ACCESS_TOKEN=
 
-# The kid to use in the token
+# The kid to use in the token for non-jaas-specific tests (though it could be a jaas key).
 #JWT_KID=
 
-# The path to the private key used for generating JWT token (.pk)
+# The path to the private key used for generating JWT token (.pk) for non-jaas-specific tests (though it could be a
+# jaas key).
 #JWT_PRIVATE_KEY_PATH=
 
 # The count of workers that execute the tests in parallel
