+++ tests/specs/helpers/jaas.ts
@@ -1,9 +1,9 @@
 import { Participant } from '../../helpers/Participant';
-import { IToken } from '../../helpers/token';
+import { IToken, ITokenOptions, generateToken } from '../../helpers/token';
 
 /**
  * Creates a new Participant and joins the MUC with the given name. The jaas-specific properties must be set as
- * environment variables: JAAS_DOMAIN and JAAS_TENANT.
+ * environment variables: JAAS_DOMAIN and JAAS_TENANT, JAAS_KID, JAAS_PRIVATE_KEY_PATH.
  *
  * @param roomName The name of the room to join, without the tenant.
  * @param instanceId This is the "name" passed to the Participant, I think it's used to match against one of the
@@ -35,4 +35,17 @@ export async function joinMuc(roomName: string, instanceId: 'p1' | 'p2' | 'p3' |
     }
 
     return newParticipant;
+}
+
+export function generateJaasToken(options: ITokenOptions): IToken {
+    if (!process.env.JAAS_PRIVATE_KEY_PATH || !process.env.JAAS_KID) {
+        throw new Error('JAAS_PRIVATE_KEY_PATH and JAAS_KID environment variables must be set');
+    }
+
+    // Don't override the keyId and keyPath if they are already set in options, allow tests to set them.
+    return generateToken({
+        ...options,
+        keyId: options.keyId || process.env.JAAS_KID,
+        keyPath: options.keyPath || process.env.JAAS_PRIVATE_KEY_PATH,
+    });
 }
